move testmain to load
